refactor(fleet): split FleetArmingRibbon into colocated subcomponents
- Extract useFleetEscapeChords hook (142 lines) — capture-phase Escape chord detection with bare-Esc and Cmd-Esc paths
- Extract useFleetRibbonFlashes hook (76 lines) — commit-flash CSS animation and window-refocus pulse
- Extract buildConfirmMessage utility (33 lines) — pending action confirm message builder + FleetConfirmActionId type
- Extract renderPaneStateBadge component (28 lines) — compact health badge for popover rows
- Extract useFleetWorktreeScope hook (63 lines) — worktree scope computation with color dedup
- Extract FleetWorktreeDots component (26 lines) — stacked worktree identity dots
- Extract FleetCountChip component (290 lines) — armed-list popover with list/picker mode state machine
- Extract SavedFleetRow component (56 lines) — saved fleet recall/delete row
- Extract SaveFleetForm component (168 lines) — inline save-fleet form with snapshot/predicate flavor toggle
- Extract SavedFleetsSection component (28 lines) — saved fleets dropdown section
- Main file shrinks from 1419 to 522 lines
